+++ guicey-spa/src/test/groovy/ru/vyarus/guicey/spa/err/SameNameTest.groovy
@@ -4,35 +4,40 @@ import io.dropwizard.Application
 import io.dropwizard.Configuration
 import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
-import org.junit.Rule
-import ru.vyarus.dropwizard.guice.test.StartupErrorRule
+import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.module.GuiceyConfigurationInfo
+import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
 import ru.vyarus.guicey.spa.SpaBundle
 import spock.lang.Specification
+
+import javax.inject.Inject
 
 /**
  * @author Vyacheslav Rusakov
  * @since 05.04.2017
  */
+@UseGuiceyApp(App)
 class SameNameTest extends Specification {
 
-    @Rule
-    StartupErrorRule rule = StartupErrorRule.create()
+    @Inject
+    GuiceyConfigurationInfo info
 
     def "Check duplicate spa names"() {
 
-        when: "starting app"
-        new App().run(['server'] as String[])
-        then: "error"
-        thrown(rule.indicatorExceptionType)
-        rule.error.contains("SPA with name 'app' is already registered")
+        expect: "duplicate registration avoided"
+        info.getInfos(SpaBundle).size() == 1
+        info.getInfos(SpaBundle)[0].registrationAttempts == 2
     }
 
     static class App extends Application<Configuration> {
 
         @Override
         void initialize(Bootstrap<Configuration> bootstrap) {
-            bootstrap.addBundle(SpaBundle.app("app", "/app", "/1").build())
-            bootstrap.addBundle(SpaBundle.app("app", "/app", "/2").build())
+            bootstrap.addBundle(GuiceBundle.builder()
+                    .bundles(
+                            SpaBundle.app("app", "/app", "/1").build(),
+                            SpaBundle.app("app", "/app", "/2").build())
+                    .build())
         }
 
         @Override